Use explicit data types to support < 64-bit systems

+++ src/config.c
@@ -255,7 +255,7 @@ static struct Config *createDefaultConfig(void)
 
 static void printUsage(const char *programName)
 {
-	int progNameLen = strlen(programName);
+	int progNameLen = (int)strlen(programName);
 	printf(
 			"%s – DHCP option injector\n"
          "Usage: %s [-df] [--forward-on-fail] [-i|-r] [-p [pid_file]] \n"
